# staticServer/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sense100 接口文档",
        "title": "sense100",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "740640431@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.5"
    },
    "host": "192.168.1.104:9494",
    "basePath": "/",
    "paths": {
...
        "/third/getWxAccessToken/{code}": {
            "get": {
                "description": "根据code获取微信accessToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Third"
                ],
                "summary": "根据code获取微信accessToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.WxAccessTokenResp"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.BaseResponseSwag": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据jsonObject",
                    "type": "string"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                }
            }
        },
        "response.WxAccessTokenResp": {
            "type": "object",
            "properties": {
                "openId": {
                    "type": "string"
                },
                "sessionKey": {
                    "type": "string"
                }
            }
        }
    }
}